Revert move session api
