+++ client/src/lib/session/hooks/useUpdateSessionState.ts
@@ -4,7 +4,7 @@ import {
   SessionState,
   LiveSession,
 } from '../../../../../shared/src/types/Session';
-import * as sessionApi from '../api/session';
+import * as sessionApi from '../../sessions/api/session';
 
 const useUpdateSessionState = (sessionId: LiveSession['id'] | undefined) => {
   const startSession = useCallback(async () => {
