+++ client/src/routes/LiveSession/ChangingRoom.tsx
@@ -44,7 +44,7 @@ import useLocalParticipant from '../../lib/daily/hooks/useLocalParticipant';
 import useUser from '../../lib/user/hooks/useUser';
 import Image from '../../lib/components/Image/Image';
 import useSubscribeToSessionIfFocused from '../../lib/session/hooks/useSusbscribeToSessionIfFocused';
-import {getSessionToken} from '../../lib/session/api/session';
+import {getSessionToken} from '../../lib/sessions/api/session';
 import useLiveSessionMetricEvents from '../../lib/session/hooks/useLiveSessionMetricEvents';
 import useCheckPermissions from '../../lib/session/hooks/useCheckPermissions';
